requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
docker
python-dotenv
pytest
sqlalchemy